+++ rollup.config.js
@@ -6,28 +6,7 @@ import uglify from 'rollup-plugin-uglify'
 export default [{
   input: './src/main.js',
   plugins: [
-    babel({
-      babelrc: false,
-      presets: [
-        [
-          '@babel/preset-env',
-          {
-            'targets': {
-              'browsers': [
-                'last 2 versions',
-                'ie >= 9'
-              ]
-            },
-            'modules': false
-          }
-        ],
-        ['@babel/preset-stage-2', {
-          'decoratorsLegacy': true
-        }]
-      ],
-      exclude: 'node_modules/**',
-      runtimeHelpers: true
-    }),
+    babel(),
     uglify(),
     filesize()
   ],
@@ -40,28 +19,7 @@ export default [{
 }, {
   input: './src/main.js',
   plugins: [
-    babel({
-      babelrc: false,
-      presets: [
-        [
-          '@babel/preset-env',
-          {
-            'targets': {
-              'browsers': [
-                'last 2 versions',
-                'ie >= 9'
-              ]
-            },
-            'modules': false
-          }
-        ],
-        ['@babel/preset-stage-2', {
-          'decoratorsLegacy': true
-        }]
-      ],
-      exclude: 'node_modules/**',
-      runtimeHelpers: true
-    }),
+    babel(),
     cleanup(),
     filesize()
   ],