+++ rollup.config.split.js
@@ -14,28 +14,7 @@ const buildEntry = () => {
     const config = {
       input: p,
       plugins: [
-        babel({
-          babelrc: false,
-          presets: [
-            [
-              '@babel/preset-env',
-              {
-                'targets': {
-                  'browsers': [
-                    'last 2 versions',
-                    'ie >= 9'
-                  ]
-                },
-                'modules': false
-              }
-            ],
-            ['@babel/preset-stage-2', {
-              'decoratorsLegacy': true
-            }]
-          ],
-          exclude: 'node_modules/**',
-          runtimeHelpers: true
-        }),
+        babel(),
         uglify(),
         filesize()
       ],