+++ tests/main.js
@@ -1,5 +1,5 @@
 import simplyValid from '../src/main'
-import test from 'ava'
+import test from 'tape'
 
 const testData = {
   zip: '11445',
@@ -11,9 +11,10 @@ test('Test String Schema', t => {
     schema: 'hasValue'
   })
 
-  t.truthy(validate)
-  t.truthy(validate('cool').isValid, 'Validation passed')
-  t.falsy(validate('').isValid, 'Empty string is not valid')
+  t.ok(validate)
+  t.ok(validate('cool').isValid, 'Validation passed')
+  t.notOk(validate('').isValid, 'Empty string is not valid')
+  t.end()
 })
 
 test('Test Object Schema', t => {
@@ -24,12 +25,13 @@ test('Test Object Schema', t => {
     }
   })
 
-  t.truthy(validate)
-  t.truthy(validate(testData).isValid, 'Object is a valid object')
-  t.falsy(validate({
+  t.ok(validate)
+  t.ok(validate(testData).isValid, 'Object is a valid object')
+  t.notOk(validate({
     zip: 'cool',
     address: '112 test St'
   }).isValid, 'Object is invalid')
+  t.end()
 })
 
 test('Test Object Schema Omitting values', t => {
@@ -39,12 +41,13 @@ test('Test Object Schema Omitting values', t => {
     }
   })
 
-  t.truthy(validate)
-  t.truthy(validate(testData).isValid, 'Object is a valid object')
-  t.falsy(validate({
+  t.ok(validate)
+  t.ok(validate(testData).isValid, 'Object is a valid object')
+  t.notOk(validate({
     zip: '',
     address: '112 test St'
   }).isValid, 'Object is invalid')
+  t.end()
 })
 
 test('Test Array Schema', t => {
@@ -52,9 +55,10 @@ test('Test Array Schema', t => {
     schema: ['isNumber', 'isPositive']
   })
 
-  t.truthy(validate)
-  t.truthy(validate('5').isValid, 'Number is valid and positive')
-  t.falsy(validate('-4').isValid, 'Number is negative and not valid')
+  t.ok(validate)
+  t.ok(validate('5').isValid, 'Number is valid and positive')
+  t.notOk(validate('-4').isValid, 'Number is negative and not valid')
+  t.end()
 })
 
 test('Test nested Object', t => {
@@ -68,14 +72,15 @@ test('Test nested Object', t => {
     }
   })
 
-  t.truthy(validate)
-  t.truthy(validate({
+  t.ok(validate)
+  t.ok(validate({
     test: '4',
     info: {
       zip: '44432',
       address: '4432 Test St'
     }
   }).isValid, 'Valid nested object')
+  t.end()
 })
 
 test('Test nested object with arrays', t => {
@@ -105,9 +110,10 @@ test('Test nested object with arrays', t => {
     }
   })
 
-  t.truthy(validate)
-  t.truthy(results.isValid, 'Is a valid object')
+  t.ok(validate)
+  t.ok(results.isValid, 'Is a valid object')
   t.is(results.story.length, 0, 'There is no story')
+  t.end()
 })
 
 test('Testing Combo Functionality', t => {
@@ -126,8 +132,9 @@ test('Testing Combo Functionality', t => {
     zip: '55555'
   })
 
-  t.truthy(results)
-  t.truthy(results.isValid)
+  t.ok(results)
+  t.ok(results.isValid)
+  t.end()
 })
 
 test('Testing Combo Functionality Mixup', t => {
@@ -146,8 +153,9 @@ test('Testing Combo Functionality Mixup', t => {
     zip: 'K1A0B1'
   })
 
-  t.truthy(results)
-  t.truthy(results.isValid)
+  t.ok(results)
+  t.ok(results.isValid)
+  t.end()
 })
 
 test('Test edge case for objects in schema', t => {
@@ -171,19 +179,38 @@ test('Test edge case for objects in schema', t => {
   }
   const results = simplyValid({ schema }, data)
 
-  t.falsy(results.isValid)
+  t.notOk(results.isValid)
   t.is(results.story[0].test, 'noLetters')
   t.is(results.story[1].test, 'noSpecials')
+  t.end()
 })
 
 test('Testing main validation for Arrays', t => {
   const results = simplyValid({ schema: ['hasNumbers', 'hasLetters'] }, ['henlo112', 'blep102'])
 
-  t.truthy(results.isValid)
+  t.ok(results.isValid)
+  t.end()
 })
 
 test('Testing main validation for Strings', t => {
   const results = simplyValid({ schema: 'hasNumbers' }, 'blep')
 
-  t.falsy(results.isValid)
+  t.notOk(results.isValid)
+  t.end()
+})
+
+test('Throws an error when the schema passed in is invalid', t => {
+  try {
+    simplyValid({ schema: [] }, 'test')
+  } catch (err) {
+    t.is(err.message, 'The schema is either invalid or one was not provided for validation')
+    t.end()
+  }
+})
+
+test('Handles if a random invalid option is passed in', t => {
+  const validator = simplyValid({ schema: ['isPositive'], hi: 'test' })
+
+  t.ok(validator(1).isValid)
+  t.end()
 })